+++ backend/src/main/java/com/sheoran/service/CouponService.java
@@ -4,10 +4,11 @@ import com.sheoran.model.Cart;
 import com.sheoran.model.Coupon;
 import com.sheoran.model.User;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface CouponService {
-    Cart applyCoupon(String code, double orderValue, User user) throws Exception;
+    Cart applyCoupon(String code, BigDecimal orderValue, User user) throws Exception;
     Cart removeCoupon(String code ,User user) throws Exception;
     Coupon findCouponById(Long id) throws Exception;
     Coupon createCoupon(Coupon coupon);
