+++ backend/src/main/java/com/sheoran/service/PaymentService.java
@@ -7,6 +7,7 @@ import com.sheoran.model.PaymentOrder;
 import com.sheoran.model.User;
 import com.stripe.exception.StripeException;
 
+import java.math.BigDecimal;
 import java.util.Set;
 
 public interface PaymentService {
@@ -15,6 +16,6 @@ public interface PaymentService {
     PaymentOrder getPaymentOrderById(Long orderId) throws Exception;
     PaymentOrder getPaymentOrderByPaymentId(String orderId) throws Exception;
     Boolean proceedPaymentOrder(PaymentOrder paymentOrder,String paymentId,String paymentLinkId) throws RazorpayException;
-    PaymentLink createRazorpayPaymentLink(User user,Long amount,Long orderId) throws RazorpayException;
-    String createStripePaymentLink(User user,Long amount,Long orderId) throws StripeException;
+    PaymentLink createRazorpayPaymentLink(User user, BigDecimal amount, Long orderId) throws RazorpayException;
+    String createStripePaymentLink(User user,BigDecimal amount,Long orderId) throws StripeException;
 }
